+++ dacite/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Vulkan 1.0.26, 1.0.27, 1.0.28, 1.0.29, 1.0.30
  - Extension `VK_EXT_validation_flags` (revision 1)
  - Extension `VK_AMD_gpu_shader_half_float` (revision 1)
+ - Extension `VK_EXT_debug_marker` (revision 3)
 
 
 ## Version 0.6.0, released on 08.07.2017